+++ staging/publishing/import-restrictions.yaml
@@ -133,6 +133,7 @@
   - k8s.io/cli-runtime
   - k8s.io/client-go
   - k8s.io/component-base
+  - k8s.io/component-helpers
   - k8s.io/kubectl
   - k8s.io/kube-openapi
   - k8s.io/metrics
