+++ staging/src/k8s.io/kubectl/pkg/cmd/auth/BUILD
@@ -1,8 +1,4 @@
-load(
-    "@io_bazel_rules_go//go:def.bzl",
-    "go_library",
-    "go_test",
-)
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
@@ -11,10 +7,9 @@ go_library(
         "cani.go",
         "reconcile.go",
     ],
-    importpath = "k8s.io/kubernetes/pkg/kubectl/cmd/auth",
-    visibility = [
-        "//build/visible_to:pkg_kubectl_cmd_auth_CONSUMERS",
-    ],
+    importmap = "k8s.io/kubernetes/vendor/k8s.io/kubectl/pkg/cmd/auth",
+    importpath = "k8s.io/kubectl/pkg/cmd/auth",
+    visibility = ["//visibility:public"],
     deps = [
         "//staging/src/k8s.io/api/authorization/v1:go_default_library",
         "//staging/src/k8s.io/api/rbac/v1:go_default_library",
@@ -44,21 +39,6 @@ go_library(
     ],
 )
 
-filegroup(
-    name = "package-srcs",
-    srcs = glob(["**"]),
-    tags = ["automanaged"],
-)
-
-filegroup(
-    name = "all-srcs",
-    srcs = [":package-srcs"],
-    tags = ["automanaged"],
-    visibility = [
-        "//build/visible_to:pkg_kubectl_cmd_auth_CONSUMERS",
-    ],
-)
-
 go_test(
     name = "go_default_test",
     srcs = ["cani_test.go"],
@@ -73,4 +53,18 @@ go_test(
         "//staging/src/k8s.io/kubectl/pkg/cmd/testing:go_default_library",
         "//staging/src/k8s.io/kubectl/pkg/scheme:go_default_library",
     ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
 )
